Make URL configurable and db creation reliable
Use `createdb` to reliably create a database.

Include LuckyRecord and use it's URL. Eventually Migrator will probably
be combine with Record. Either that or I may extract a
PostgresURL library for creating URLs so that Migrator doesn't need to
rely on Record. For now this is fine though since most will probably use
Migrator and Record together