Execute migration statements in order (#87)
Closes #84 

When I changed the `Migration` to execute statements in a transaction, I collected them in `@prepared_statements` then executed them in order. But the `Migrations#execute` method executed the statement outside of the transaction as soon as it was called. This PR fixes that.

The one catch is that I had to restrict the `execute` method to strings. I could have allows it to accept `*args`, but that requires building the statement by escaping the the arguments manually, or significantly changing the `@prepared_statements` array and `execute_in_transaction` to accommodate it.

Lastly, I wrote a spec to catch this error, but it's dependent on `LuckyRecord::table_columns` which isn't the `LuckyRecord` version used by `LuckyMigrator`. Should I update the dependencies in a separate PR?